+++ tests/env/plugin/test_interface.py
@@ -39,7 +39,7 @@ class TestEnvVars:
         config = {'project': {'name': 'my_app', 'version': '0.0.1'}}
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.env_vars == environment.env_vars == {AppEnvVars.ENV_ACTIVE: 'default'}
@@ -51,7 +51,7 @@ class TestEnvVars:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(TypeError, match='Field `tool.hatch.envs.default.env-vars` must be a mapping'):
@@ -64,7 +64,7 @@ class TestEnvVars:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(
@@ -79,7 +79,7 @@ class TestEnvVars:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.env_vars == {AppEnvVars.ENV_ACTIVE: 'default', 'foo': 'bar'}
@@ -90,7 +90,7 @@ class TestEnvInclude:
         config = {'project': {'name': 'my_app', 'version': '0.0.1'}}
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.env_include == environment.env_include == []
@@ -102,7 +102,7 @@ class TestEnvInclude:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(TypeError, match='Field `tool.hatch.envs.default.env-include` must be an array'):
@@ -115,7 +115,7 @@ class TestEnvInclude:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(
@@ -130,7 +130,7 @@ class TestEnvInclude:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.env_include == ['HATCH_BUILD_*', 'FOO*']
@@ -141,7 +141,7 @@ class TestEnvExclude:
         config = {'project': {'name': 'my_app', 'version': '0.0.1'}}
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.env_exclude == environment.env_exclude == []
@@ -153,7 +153,7 @@ class TestEnvExclude:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(TypeError, match='Field `tool.hatch.envs.default.env-exclude` must be an array'):
@@ -166,7 +166,7 @@ class TestEnvExclude:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(
@@ -181,7 +181,7 @@ class TestEnvExclude:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.env_exclude == ['FOO*']
@@ -192,7 +192,7 @@ class TestPlatforms:
         config = {'project': {'name': 'my_app', 'version': '0.0.1'}}
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.platforms == environment.platforms == []
@@ -204,7 +204,7 @@ class TestPlatforms:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(TypeError, match='Field `tool.hatch.envs.default.platforms` must be an array'):
@@ -217,7 +217,7 @@ class TestPlatforms:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(
@@ -232,7 +232,7 @@ class TestPlatforms:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.platforms == ['macos']
@@ -243,7 +243,7 @@ class TestSkipInstall:
         config = {'project': {'name': 'my_app', 'version': '0.0.1'}}
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.skip_install is environment.skip_install is False
@@ -255,7 +255,7 @@ class TestSkipInstall:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(TypeError, match='Field `tool.hatch.envs.default.skip-install` must be a boolean'):
@@ -268,7 +268,7 @@ class TestSkipInstall:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.skip_install is True
@@ -279,7 +279,7 @@ class TestDevMode:
         config = {'project': {'name': 'my_app', 'version': '0.0.1'}}
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.dev_mode is environment.dev_mode is True
@@ -291,7 +291,7 @@ class TestDevMode:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(TypeError, match='Field `tool.hatch.envs.default.dev-mode` must be a boolean'):
@@ -304,7 +304,7 @@ class TestDevMode:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.dev_mode is False
@@ -315,7 +315,7 @@ class TestFeatures:
         config = {'project': {'name': 'my_app', 'version': '0.0.1'}}
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.features == environment.features == []
@@ -327,7 +327,7 @@ class TestFeatures:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(TypeError, match='Field `tool.hatch.envs.default.features` must be an array of strings'):
@@ -340,7 +340,7 @@ class TestFeatures:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.features == ['baz', 'foo-bar']
@@ -352,7 +352,7 @@ class TestFeatures:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(TypeError, match='Feature #1 of field `tool.hatch.envs.default.features` must be a string'):
@@ -365,7 +365,7 @@ class TestFeatures:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(
@@ -380,7 +380,7 @@ class TestFeatures:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(
@@ -398,7 +398,7 @@ class TestDescription:
         config = {'project': {'name': 'my_app', 'version': '0.0.1'}}
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.description == environment.description == ''
@@ -410,7 +410,7 @@ class TestDescription:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(TypeError, match='Field `tool.hatch.envs.default.description` must be a string'):
@@ -424,7 +424,7 @@ class TestDescription:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.description is description
@@ -435,7 +435,7 @@ class TestDependencies:
         config = {'project': {'name': 'my_app', 'version': '0.0.1', 'dependencies': ['dep1']}}
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.dependencies == environment.dependencies == ['dep1']
@@ -448,7 +448,7 @@ class TestDependencies:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(TypeError, match='Field `tool.hatch.envs.default.dependencies` must be an array'):
@@ -461,7 +461,7 @@ class TestDependencies:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(
@@ -476,7 +476,7 @@ class TestDependencies:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(
@@ -491,7 +491,7 @@ class TestDependencies:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(TypeError, match='Field `tool.hatch.envs.default.extra-dependencies` must be an array'):
@@ -504,7 +504,7 @@ class TestDependencies:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(
@@ -519,7 +519,7 @@ class TestDependencies:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(
@@ -534,7 +534,7 @@ class TestDependencies:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.dependencies == ['dep2', 'dep3', 'dep1']
@@ -548,7 +548,7 @@ class TestDependencies:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         normalized_path = str(isolation).replace('\\', '/')
@@ -567,7 +567,7 @@ class TestDependencies:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.dependencies == ['dep2', 'dep3']
@@ -583,7 +583,7 @@ class TestDependencies:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.dependencies == ['dep2', 'dep3']
@@ -597,7 +597,7 @@ class TestScripts:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(TypeError, match='Field `tool.hatch.envs.default.scripts` must be a table'):
@@ -610,7 +610,7 @@ class TestScripts:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(
@@ -622,7 +622,7 @@ class TestScripts:
         config = {'project': {'name': 'my_app', 'version': '0.0.1'}}
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.scripts == environment.scripts == {}
@@ -635,7 +635,7 @@ class TestScripts:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.scripts == {'foo': ['command1'], 'bar': ['command2']}
@@ -648,7 +648,7 @@ class TestScripts:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.scripts == {'foo': ['command1'], 'bar': ['command3', 'command2']}
@@ -660,7 +660,7 @@ class TestScripts:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(
@@ -675,7 +675,7 @@ class TestScripts:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(
@@ -691,7 +691,7 @@ class TestScripts:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.scripts == {'foo': ['command1'], 'bar': ['command3', 'command1']}
@@ -708,7 +708,7 @@ class TestScripts:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.scripts == {
@@ -729,7 +729,7 @@ class TestScripts:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.scripts == {
@@ -745,7 +745,7 @@ class TestScripts:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(
@@ -760,7 +760,7 @@ class TestPreInstallCommands:
         config = {'project': {'name': 'my_app', 'version': '0.0.1'}}
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.pre_install_commands == environment.pre_install_commands == []
@@ -772,7 +772,7 @@ class TestPreInstallCommands:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(TypeError, match='Field `tool.hatch.envs.default.pre-install-commands` must be an array'):
@@ -785,7 +785,7 @@ class TestPreInstallCommands:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(
@@ -800,7 +800,7 @@ class TestPreInstallCommands:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.pre_install_commands == ['baz test']
@@ -811,7 +811,7 @@ class TestPostInstallCommands:
         config = {'project': {'name': 'my_app', 'version': '0.0.1'}}
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.post_install_commands == environment.post_install_commands == []
@@ -823,7 +823,7 @@ class TestPostInstallCommands:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(TypeError, match='Field `tool.hatch.envs.default.post-install-commands` must be an array'):
@@ -836,7 +836,7 @@ class TestPostInstallCommands:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(
@@ -851,7 +851,7 @@ class TestPostInstallCommands:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.post_install_commands == ['baz test']
@@ -862,7 +862,7 @@ class TestEnvVarOption:
         config = {'project': {'name': 'my_app', 'version': '0.0.1'}}
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert environment.get_env_var_option('foo') == ''
@@ -871,7 +871,7 @@ class TestEnvVarOption:
         config = {'project': {'name': 'my_app', 'version': '0.0.1'}}
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with EnvVars({'HATCH_ENV_TYPE_MOCK_FOO': 'bar'}):
@@ -886,7 +886,7 @@ class TestContextFormatting:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert list(environment.expand_command('foo')) == ['command default']
@@ -898,7 +898,7 @@ class TestContextFormatting:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert list(environment.expand_command('foo')) == ['command mock']
@@ -910,7 +910,7 @@ class TestContextFormatting:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 9000
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 9000
         )
 
         assert list(environment.expand_command('foo')) == ['command -v=9000']
@@ -922,7 +922,7 @@ class TestContextFormatting:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         with pytest.raises(ValueError, match='Unknown verbosity modifier: bar'):
@@ -935,10 +935,10 @@ class TestContextFormatting:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
-        with pytest.raises(ValueError, match='Verbosity flag adjustment must be an integer: -1.0'):
+        with pytest.raises(TypeError, match='Verbosity flag adjustment must be an integer: -1.0'):
             next(environment.expand_command('foo'))
 
     @pytest.mark.parametrize(
@@ -962,7 +962,7 @@ class TestContextFormatting:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, verbosity
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, verbosity
         )
 
         assert list(environment.expand_command('foo')) == [command]
@@ -988,7 +988,7 @@ class TestContextFormatting:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert list(environment.expand_command('foo')) == [command]
@@ -1000,7 +1000,7 @@ class TestContextFormatting:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert list(environment.expand_command('foo')) == ['command']
@@ -1012,7 +1012,7 @@ class TestContextFormatting:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert list(environment.expand_command('foo')) == ['command  -bar > /dev/null']
@@ -1024,7 +1024,57 @@ class TestContextFormatting:
         }
         project = Project(isolation, config=config)
         environment = MockEnvironment(
-            isolation, project.metadata, 'default', project.config.envs['default'], data_dir, platform, 0
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
         )
 
         assert list(environment.expand_command('foo baz')) == ['command baz']
+
+    def test_matrix_no_selection(self, isolation, data_dir, platform):
+        config = {
+            'project': {'name': 'my_app', 'version': '0.0.1'},
+            'tool': {'hatch': {'envs': {'default': {'dependencies': ['pkg=={matrix}']}}}},
+        }
+        project = Project(isolation, config=config)
+        environment = MockEnvironment(
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
+        )
+
+        with pytest.raises(ValueError, match='The `matrix` context formatting field requires a modifier'):
+            _ = environment.dependencies
+
+    def test_matrix_no_default(self, isolation, data_dir, platform):
+        config = {
+            'project': {'name': 'my_app', 'version': '0.0.1'},
+            'tool': {'hatch': {'envs': {'default': {'dependencies': ['pkg=={matrix:bar}']}}}},
+        }
+        project = Project(isolation, config=config)
+        environment = MockEnvironment(
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
+        )
+
+        with pytest.raises(ValueError, match='Nonexistent matrix variable must set a default: bar'):
+            _ = environment.dependencies
+
+    def test_matrix_default(self, isolation, data_dir, platform):
+        config = {
+            'project': {'name': 'my_app', 'version': '0.0.1'},
+            'tool': {'hatch': {'envs': {'default': {'dependencies': ['pkg=={matrix:bar:9000}']}}}},
+        }
+        project = Project(isolation, config=config)
+        environment = MockEnvironment(
+            isolation, project.metadata, 'default', project.config.envs['default'], {}, data_dir, platform, 0
+        )
+
+        assert environment.dependencies == ['pkg==9000']
+
+    def test_matrix_default_override(self, isolation, data_dir, platform):
+        config = {
+            'project': {'name': 'my_app', 'version': '0.0.1'},
+            'tool': {'hatch': {'envs': {'default': {'dependencies': ['pkg=={matrix:bar:baz}']}}}},
+        }
+        project = Project(isolation, config=config)
+        environment = MockEnvironment(
+            isolation, project.metadata, 'default', project.config.envs['default'], {'bar': '42'}, data_dir, platform, 0
+        )
+
+        assert environment.dependencies == ['pkg==42']